+++ .gitignore
@@ -97,3 +97,5 @@ pyvenv.cfg
 
 .idea/
 
+logsetup.json
+**/*~
