+++ roku2mqtt/__main__.py
@@ -1,4 +1,4 @@
-from roku.roku import main
+from roku2mqtt.roku import main
 
 if __name__ == '__main__':
     main()
